### Benchmark Comparison

#### Key Changes:
1. **Small Inputs**:
   - `SmallASCII`: Slightly slower (137.1 ns → 145.2 ns).
   - `SmallUnicode`: Slower (192.1 ns → 242.1 ns), added allocations (0 B/op → 64 B/op).

2. **Large Inputs**:
   - `LargeASCII`: Slightly faster (47575 ns → 46160 ns), reduced allocations (2016 B/op → 1024 B/op).
   - `LargeUnicode`: No significant change.

3. **Edge Cases**:
   - `IdenticalStrings`, `EmptyStrings`, `OneEmpty`: Dramatically faster (ns → ~2 ns), no allocations.

#### Summary:
- **Gains**: Major speedups and zero allocations for edge cases.
- **Losses**: Minor regressions in small inputs, especially Unicode.
- **Trade-offs**: Optimizations favor edge cases but slightly impact small inputs.

before:
```
➜  deeply git:(master) ✗ go test -bench .
goos: darwin
goarch: arm64
pkg: github.com/gripmock/deeply
cpu: Apple M1 Pro
BenchmarkDistance_SmallASCII-8           7784780               137.1 ns/op             0 B/op          0 allocs/op
BenchmarkDistance_SmallUnicode-8         6317986               192.1 ns/op             0 B/op          0 allocs/op
BenchmarkDistance_LargeASCII-8             25010             47575 ns/op            2016 B/op          3 allocs/op
BenchmarkDistance_LargeUnicode-8           24993             47875 ns/op            2016 B/op          3 allocs/op
BenchmarkDistance_IdenticalStrings-8       92086             12959 ns/op             512 B/op          2 allocs/op
BenchmarkDistance_EmptyStrings-8        66816074                17.91 ns/op            0 B/op          0 allocs/op
BenchmarkDistance_OneEmpty-8             9258082               123.5 ns/op           256 B/op          1 allocs/op
PASS
ok      github.com/gripmock/deeply      10.129s
```

after:
```
➜  deeply git:(optimize) ✗ go test -bench .
goos: darwin
goarch: arm64
pkg: github.com/gripmock/deeply
cpu: Apple M1 Pro
BenchmarkDistance_SmallASCII-8           7686594               145.2 ns/op             0 B/op          0 allocs/op
BenchmarkDistance_SmallUnicode-8         5010283               242.1 ns/op            64 B/op          1 allocs/op
BenchmarkDistance_LargeASCII-8             25712             46160 ns/op            1024 B/op          1 allocs/op
BenchmarkDistance_LargeUnicode-8           25251             47510 ns/op            2016 B/op          3 allocs/op
BenchmarkDistance_IdenticalStrings-8    478831730                2.495 ns/op           0 B/op          0 allocs/op
BenchmarkDistance_EmptyStrings-8        479017215                2.529 ns/op           0 B/op          0 allocs/op
BenchmarkDistance_OneEmpty-8            580798672                2.058 ns/op           0 B/op          0 allocs/op
PASS
ok      github.com/gripmock/deeply      10.968s
```